+++ module/utility/defense.mjs
@@ -1,3 +1,5 @@
+import { getPowerInfo } from "../utility/util.mjs";
+
 function determineDefense(targetActor, attackItem, options) {
     if (!attackItem.findModsByXmlid) {
         console.error("Invalid attackItem", attackItem);
@@ -289,8 +291,30 @@ function determineDefense(targetActor, attackItem, options) {
             game.settings.get("hero6efoundryvttv2", "knockback") &&
             attackItem.system.knockbackMultiplier
         ) {
-            if (["KBRESISTANCE", "DENSITYINCREASE", "GROWTH"].includes(xmlid)) {
+            if (["KBRESISTANCE", "DENSITYINCREASE"].includes(xmlid)) {
                 let _value = value * (targetActor.system.is5e ? 1 : 2);
+                knockbackResistance += _value;
+                defenseTags.push({
+                    value: _value,
+                    name: "KB" + tagXmlids,
+                    title: activeDefense.name,
+                });
+            }
+
+            if (["GROWTH"].includes(xmlid)) {
+                const configPowerInfo = getPowerInfo({ item: activeDefense });
+                const details = configPowerInfo?.details(activeDefense) || {};
+                let _value = details.kb; //value * (targetActor.system.is5e ? 1 : 6);
+                knockbackResistance += _value;
+                defenseTags.push({
+                    value: _value,
+                    name: "KB" + tagXmlids,
+                    title: activeDefense.name,
+                });
+            }
+
+            if (["SHRINKING"].includes(xmlid)) {
+                let _value = -value * (targetActor.system.is5e ? 1 : 6);
                 knockbackResistance += _value;
                 defenseTags.push({
                     value: _value,