+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Presence attack can now do 1/2 dice and 1/2 dice modification increments. [#943](https://github.com/dmdorman/hero6e-foundryvtt/issues/943)
 - Attacks now have auto success (3) and auto failure (18). [#562](https://github.com/dmdorman/hero6e-foundryvtt/issues/562)
 - Warning is now given when attacking with no target selected. [#927](https://github.com/dmdorman/hero6e-foundryvtt/issues/927)
+- Fixed issue where SHRINKING failed to toggle on/off. [#936](https://github.com/dmdorman/hero6e-foundryvtt/issues/936)
+- Improved SHRINKING and GROWTH support.
 
 ## Version 3.0.66 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)